requirements.txt
streamlit==1.13.0
pandas==1.5.0
plotly.express==5.10.0
Pillow==9.2.0